threading: Fix hang on upload completion
There was a small race condition in the fetching thread. The fetching
thread would check the download thread, see that it's still alive, and
block forever trying to fetch an item from the queue. Meanwhile, the
download thread has completed, and exits without adding anything to the
queue, resulting in deadlock.

Add a timeout to the fetch that will allow it to detect when the
download thread has completed.